pyproject.toml
[tool.poetry]
name = "python-grpc-example"
version = "0.1.0"
description = ""
authors = ["tomato2718 <qweasdzxc030@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.*"
grpcio = "1.64.1"


[tool.poetry.group.dev.dependencies]
grpcio-tools = "1.64.1"


[tool.poetry.group.server.dependencies]
uvloop = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
